pyproject.toml
[project]
name = "cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "edge-tts>=7.2.7",
    "google-generativeai>=0.5.0",
    "lumaai>=0.0.1",
    "moviepy<2.0.0",
    "openai>=1.0.0",
    "pypinyin>=0.55.0",
    "pyyaml>=6.0",
    "requests>=2.32.5",
    "volcengine>=1.0.212",
    "volcengine-python-sdk>=5.0.3",
]
